requirements.txt
asyncio
bittensor
bittensor-wallet
python-dotenv